+++ client-app/core/composables/useEnvironmentName.ts
@@ -1,11 +1,4 @@
 import { useThemeContext } from "@/core/composables/useThemeContext";
-import { COLORS } from "@/core/constants";
-
-const BADGE_COLORS: Record<string, VcBadgeColorType> = {
-  dev: COLORS.secondary,
-  qa: COLORS.info,
-  demo: COLORS.success,
-};
 
 const ENVIRONMENTS_TO_IGNORE = ["prod", "production"];
 
@@ -16,12 +9,10 @@ export function useEnvironmentName() {
     storeSettings: { environmentName },
   } = themeContext.value;
 
-  const badgeColor: VcBadgeColorType = BADGE_COLORS[environmentName?.toLowerCase()] ?? COLORS.neutral;
-  const shouldShowBadge = !!environmentName && !ENVIRONMENTS_TO_IGNORE.includes(environmentName?.toLowerCase());
+  const isIgnored = !environmentName || ENVIRONMENTS_TO_IGNORE.includes(environmentName?.toLowerCase());
 
   return {
-    badgeColor,
     environmentName,
-    shouldShowBadge,
+    isIgnored,
   };
 }
